removed Okta mgmt in demo, etc
