+++ backend/Dockerfile.prod
@@ -16,7 +16,7 @@ COPY --chown=gradle:gradle ./config ./config
 COPY --chown=gradle:gradle ./src ./src
 COPY --chown=gradle:gradle gradle.properties gradle.properties
 
-RUN gradle --no-daemon classes testClasses assemble
+RUN gradle --no-daemon classes assemble
 
 FROM openjdk:11-jre-slim
 EXPOSE 8080
