+++ backend/src/main/java/gov/cdc/usds/simplereport/service/OrganizationInitializingService.java
@@ -23,6 +23,8 @@ import gov.cdc.usds.simplereport.db.repository.FacilityRepository;
 import gov.cdc.usds.simplereport.db.repository.OrganizationRepository;
 import gov.cdc.usds.simplereport.db.repository.ProviderRepository;
 
+import com.okta.sdk.resource.ResourceException;
+
 @Service
 @Transactional
 public class OrganizationInitializingService {
@@ -47,6 +49,7 @@ public class OrganizationInitializingService {
 		Organization emptyOrg = _props.getOrganization();
 		Optional<Organization> probe = _orgRepo.findByExternalId(emptyOrg.getExternalId());
 		if (probe.isPresent()) {
+			initOktaOrg(probe.get());
 			return; // one and done
 		}
 		Provider savedProvider = _providerRepo.save(_props.getProvider());
@@ -66,20 +69,23 @@ public class OrganizationInitializingService {
 		DeviceType defaultDeviceType = configured.get(0);
 		LOG.info("Creating organization {}", emptyOrg.getOrganizationName());
 		Organization realOrg = _orgRepo.save(emptyOrg);
-		_oktaService.createOrganization(emptyOrg.getOrganizationName(), emptyOrg.getExternalId());
+		// in the unlikely event DB and Okta fall out of sync
+		initOktaOrg(realOrg);
 		Facility defaultFacility = _props.getFacility().makeRealFacility(realOrg, savedProvider, defaultDeviceType, configured);
 		LOG.info("Creating facility {} with {} devices configured", defaultFacility.getFacilityName(), configured.size());
 		_facilityRepo.save(defaultFacility);
 	}
 
+	private void initOktaOrg(Organization org) {
+		try {
+			LOG.info("Creating organization {} in Okta", org.getOrganizationName());
+			_oktaService.createOrganization(org.getOrganizationName(), org.getExternalId());
+		} catch (ResourceException e) {
+			LOG.info("Organization {} already exists in Okta", org.getOrganizationName());
+		}
+	}
+
 	public String getDefaultOrganizationId() {
 		return _props.getOrganization().getExternalId();
 	}
-
-	@PreDestroy
-    public void preDestroy() {
-		Organization emptyOrg = _props.getOrganization();
-		LOG.info("Removing organization {} from Okta", emptyOrg.getOrganizationName());
-        _oktaService.deleteOrganization(emptyOrg.getExternalId());
-    }
 }
