+++ backend/src/main/resources/application.yaml
@@ -45,9 +45,10 @@ simple-report:
     max-csv-rows: 999
     api-key: ${DATAHUB_API_KEY:MISSING}
     secret-slack-notify-webhook-url: ${SECRET_SLACK_NOTIFY_WEBHOOK_URL:MISSING}
-  authorization:
-    role-claim: prod_roles
-    role-prefix: "SR-PROD-TENANT:"
+  authorization:    
+    # these are overridden by application-dev, -prod, -test, etc
+    role-claim: dev_roles
+    role-prefix: "SR-DEV-TENANT:"
 logging:
   pattern:
     console: "%clr(%d{yyyy-MM-dd HH:mm:ss.SSS}){faint} %clr(%5p) %clr(${PID:- }){magenta} %clr(---){faint} %clr([%15.15t]){faint} [Query: %X{graphql-query}] %clr(%-40.40logger{39}){cyan} %clr(:){faint} %m%n%wEx"
