+++ backend/src/test/java/gov/cdc/usds/simplereport/api/BaseApiTest.java
@@ -23,8 +23,8 @@ import com.graphql.spring.boot.test.GraphQLTestTemplate;
 import gov.cdc.usds.simplereport.config.authorization.OrganizationRole;
 import gov.cdc.usds.simplereport.config.authorization.OrganizationRoles;
 import gov.cdc.usds.simplereport.service.AuthorizationService;
-import gov.cdc.usds.simplereport.service.OktaService;
 import gov.cdc.usds.simplereport.service.OrganizationInitializingService;
+import gov.cdc.usds.simplereport.service.OktaService;
 import gov.cdc.usds.simplereport.test_util.DbTruncator;
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
@@ -44,8 +44,8 @@ public abstract class BaseApiTest {
     @MockBean
     protected AuthorizationService _authService;
 
-    @MockBean
-    protected OktaService _oktaService;
+    @Autowired
+    private OktaService _oktaService;
 
     protected void truncateDb() {
         _truncator.truncateAll();
@@ -62,6 +62,7 @@ public abstract class BaseApiTest {
     @AfterEach
     public void cleanup() {
         truncateDb();
+        _oktaService.deleteOrganization(_initService.getDefaultOrganizationId());
     }
 
     /**
